collect-logs: run in two stages
We should run and publish a minimal sets of files even without any host,
when we failed the run before inventory generation.

This change separates the collection step that runs on all hosts except
localhost, and the rest running on localhost. Running on localhost
always succeeds, even with an empty inventory.

Also add a log environment file for local collect-logs.sh runs that
does not upload logs.

Change-Id: I48d07d42be879026fb80afd73835484770006f85