validate-tempest: use local_working_dir var
Because of the use of the PWD environment variable, it's quite tricky to
reuse validate-tempest out of the tripleo-quickstart context. This patch
replaces the variable with local_working_dir, which is already widely
used for this purpose.
The patch also adjust collect-logs's artcl_full_artifact_url for the same
reason.

Change-Id: I3647952dd298df6cb5f18cea45d7a30b7268ec85