Add upgrade and update logs to logstash indexing
Add log files from upgrades to indexing into logstash

This will add all files with a .log extension/suffix
containing "upgrade" or "update" in the filename to
the list of files for logstash indexing.

As some of the logging in tripleo-upgrade contains
release names and is in flux, this should ensure that
future filenames (with future release names embedded)
as well as newly added logfiles (planned) are indexed.

Change-Id: Iea2f38112ad2ecfb0bc0dae149b3a12fdd937d5f