Introduce zuul.projects and executed pre-run playbooks for reproducer
This uses the zuul-variables file for zuul.projects and the json console
log written by the depends on, for the playbooks. This is tracked by the
ci team with [1]. These are made available to but not yet used in the
reproducer (that will come later, see the story linked in [1]).

[1] https://tree.taiga.io/project/tripleo-ci-board/task/270
Depends-On: https://review.openstack.org/615191
Co-Authored-By: Ronelle Landy <rlandy@redhat.com>
Change-Id: I64923cfd75e697b98507be1ff398c14654108ddf