Reproducer script: run only on job started by zuul
The reproducer script is assuming a job was started by zuul. In some instances
we are still running jobs from platform different than zuul like ci.centos. In
those instances the log collection fails because the reproducer script creation
misses a lot of files
This patch includes that role only under the condition that zuul variable is
defined.

Change-Id: I01ec134f85943a8332832bdc6caf138d978c1c37