Split the collect logs task
This split by type in the collect logs is to enable thirdy party who
want's to use collect logs, to choose which type of log they want to
collect. For example, some don't have interest in dstat, or only cares
about openstack logs, or is sure that there isn't container running.

Change-Id: I8e681eef35084a81c7530f51cd6fae0187c6116d