Use ansible_python.executable to determine pip module
In collect-logs role, we use pip to collect all the pip installed
modules but on system pip2 or pip3 is installed or it is not there
leading to failure like command not found.

Using ansible_python.executable -m to get the pip module and use
the same to collect it will solve py2 and py3 issue.

For getting ansible_python.executable, we need to gather python
facts.

Change-Id: I8c8acd71869bbc2292ee935b931d9ba43656f8f9
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>