Enable compression option and tar on xfer prior to publish
When collecting logs of many nodes and storing them on
the primary nodepool node or slave it makes sense
to compress the files, tar and send. Pretty standard
logic there.   In upstream the default remains not
to compress prior to tar or prior to publish. There
is a sunk time cost in compressing and uncompressing.

Note: this removes the optional logic to compress
based on file size as it make more sense to compress
all the files prior to the tar and xfer to the
primary node / slave for faster results.

Signed off by Wes Hayutin.

Close-Bug: #1862612
Change-Id: I550e5c5ead7f65fba184f311a3a4b1c64a42a6d2