Add ignore_errors in the ir plugin playbook
Plugin's main playbook must continue in execution, no matter what,
so the whole playbook should ignore errors.

More background info: not all tasks of the role are ignoring errors,
which has caused an issue in the situation when the networking got
broken suddenly and ansible lost connection to the nodes while executing
the role -> after the connection got lost, tasks started failing and when
one of the tasks not ignoring errors failed, the whole playbook failed
which resulted in not calling the other half of the plugin's playbook,
which is responsible for publishing ...

Anyway, the role is supposed to gather as much as possible, so no matter
errors, all the tasks within it should be executed.

The patch adds ignore_errors in ir plugin's playbook, which shields
ir plugin from any errors and provides execution of all the tasks when
running the as infrared plugin.

Change-Id: Ib70a4b72fbed2eda730eb279e175107846fffeec