Replace top with vmstat
When top is run on centos-8 it creates .config/procps/ However, when
top is run in containers it creates .config/procps/ in every mount. That
creates a problem for mysql/mariadb as it tries to treat .config as
database

MariaDB [(none)]> show databases;
+--------------------+
| Database           |
+--------------------+
| #mysql50#.config   |

When something tries to restart mysql_bundle it returns non zero return
code as it reports the issue that #mysql50#.config which fails
standalone upgrade process. This patch replaces top with vmstat -s and
ps axfo equivalent.

Change-Id: I331e17d6b95b021647c9f2c36ad104da75626fa1