Fix rsync mode filters to allow *.log files
With the current defaults, there is no *.log files matching the
templated rsync filter from the nested dirs. While the nested dirs
itself are allowed to the filter, but have nothing matching because of
that.

Allow *.log files into that filter in order to collect logs properly.
Additionally, exclude redundant or not much useful paths:
  - overlay*
  - root
  - console*primary.log
  - anaconda*
since the former increases the size of collected logs by ~35M, while
beeing mostly useless. And the others contain redundant things
or junk.

Depends-On: https://review.opendev.org/760335
Change-Id: I3548e3b89c0654acca68c1dfe7a266172d3f2621
Closes-bug: #1901508
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>