Fix extra/podman/containers sometimes missing
There is bug that sometimes /var/log/extra/podman/containers
is completely missing for some nodes (while it may be there for others).

It's caused by timing of parallel-container-log-generation+sync
and generating the list for files to collect using find.

If container logs are not generated yet or in progress,
find may miss these files and so later rsync will not collect them.

Fix seems to be simple - to move the step Wait-for-containers-logs
just before the execution of find.

Closes-Bug: #1905991
Change-Id: Ib564a4f4271108015a6b54669789fffdb43d19c4