Find correct python interpreter
If /usr/bin/python is not available when trying to collect
the logs, then none are collected. In this patch we ensure
that alternative python interpreters are considered when
attempting the log collection. The interpreters are
considered in order of preference, with the system python
being the last and the default when the host is unreachable.

This resolves the issue when infrared has an interpreter
set in the inventory, but it's the wrong one due to the
host recently being upgraded from RHEL7 to RHEL8.

Change-Id: I62ec9b5ed1978806c82a973bb901d527c731d3a9