Add 2 minute timeout for repoquery
We encountered an issue when repoquery took several minutes
which has lead to timeouts and unfinished log collections.
To avoid that this commit adds a timeout which is basically
a fail-safe from collect_logs point of view.

The patch replaces for loop for record_available_packages command
by a single command.

Change-Id: Ie3007414aac14db47696fca62b07e1efa4e1de16