Add new script for SELinux consolidation
In order to make sealert faster and avoid any timeout, let's consolidate
the found AVC.

In order to do so, we can ignore the MCS level of the scontext and
tcontext. We also ensure we capture all the potential actions (read,
write and so on) as well as the tclass (file, socket, and so on).

This allows to get a cleaner, shorter list of AVC, making the overall
run faster for sealert.

It also starts sealert without dbus service as stand alone app.

Note: no shebang nor executable script due to the ansible sanity
check[1]. Perl isn't known...

[1] https://github.com/ansible/ansible/blob/devel/test/lib/ansible_test/_util/controller/sanity/code-smell/shebang.py

Related-Bug: #1946763
Change-Id: I8801137b88ce1a09e2d7ba4a5364e6e19704b6de