Fix molecule job
This adds the possibility to sova uses a local file instead of
downloading from github.
This is required because the verification searchs for a specific
failure, that recently was updated in the sova config file, and
the test started to fail. So, now, we use a static sova config
file to be used by the molecule, avoiding these kind of failures
happen in the future.

Change-Id: I2530188f1795a39ec6ec7e43c9ba118349c08541