Some linting, and correcting deps for sanity
ansible-test seems to take the latest ansible-core by default, even if
we push upper/lower constraints in the requirements.txt or
test-requirements.txt.

Pushing the constraint directly in the tox.ini, for the "sanity"
testenv, seems to do the trick, as we're testing against an ansible
version we actually support, instead of the latest (2.13 today).

This patch also takes the opportunity to remove unwanted warnings
(ansible.cfg deprecated options), as well as ensuring the ignored
content for `ansible-test' doesn't raise any warning.

Closes-Bug: #1976251
Change-Id: I596aecd71ebe7d8a747ab8e6d05766f70dd78f1d