Get a dedicated, well formatted file for dropped packets
With the current[1] configuration, nft logs end in journald. In order
to ensure we get a readable output, let's just grab the important
lines and create a dedicated file.

This will NOT work for iptables backend, since nftables is configured
with some specificities we can't set in iptables.

Since the log jump is at the very end of the chain, right before the
drop, anything logged is a packet that doesn't match any rule; we can
therefore understand them as "dropped packets", and would be able to
understand the actual reason, and add potential missing rules if
needed.

[1] https://review.opendev.org/c/openstack/tripleo-ansible/+/850222

Change-Id: I36d2fe7340c525c0438efc80bf73a103d90aa44d