Fix ansible sanity and linters jobs
The ansible sanity job was failing because the new node
that runs the job was updated and no longer have python 3.8.
In order to use python 3.10 you need to update ansible-core
to at least 2.13, doing so, it also requires that we add
the ignore-2.13.txt, and update the zuul.d/layout.yaml.
The pylint complains then about use-a-generator, so updating
sova to fix the issue, once the PEP0289 [1] is from 2003,
it should work on all python versions.

The linters job started to fails because gitlab requires now
authentication to clone the flake8 repo, so moving to flake8
repo on github.

1 - https://peps.python.org/pep-0289/

Change-Id: I1370d41b0c4e057c1c65cfd75ff945d648696771